# docs/swagger.yaml
basePath: /
definitions:
  handlers.APIErrorResponse:
    properties:
      error:
        type: string
    type: object
  handlers.APISearchResponse:
    properties:
      search_results:
        items:
          $ref: '#/definitions/handlers.SearchResult'
        type: array
    type: object
  handlers.SearchResult:
    properties:
      description:
        type: string
      id:
        type: integer
      language:
        type: string
      title:
        type: string
      url:
        type: string
    type: object
  handlers.WeatherAPIResponse:
    properties:
      forecast:
        $ref: '#/definitions/handlers.WeatherForecast'
      location:
        $ref: '#/definitions/handlers.WeatherLocation'
    type: object
  handlers.WeatherForecast:
    properties:
      step:
        type: string
      temperature:
        type: number
      wind_direction:
        type: number
      wind_speed:
        type: number
    type: object
  handlers.WeatherLocation:
    properties:
      latitude:
        type: number
      longitude:
        type: number
    type: object
info:
  contact: {}
  description: 'API for the WhoKnows web app: session auth, search content, weather
    forecast, and health/readiness probes.'
  title: WhoKnows API
  version: 0.1.0
paths:
  /api/login:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Authenticate a user and start a session.
      parameters:
      - description: Username
        in: formData
        name: username
        required: true
        type: string
      - description: Password
        in: formData
        name: password
        required: true
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: Rendered login form with errors
          schema:
            type: string
        "302":
          description: Redirect to home page
          schema:
            type: string
      summary: User login
      tags:
      - Auth
  /api/logout:
    post:
      description: Clear the user session and redirect home.
      produces:
      - text/html
      responses:
        "302":
          description: Redirect to home page
          schema:
            type: string
      security:
      - sessionAuth: []
      summary: Logout user
      tags:
      - Auth
  /api/register:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Create a new user account.
      parameters:
      - description: Username
        in: formData
        name: username
        required: true
        type: string
      - description: Email address
        in: formData
        name: email
        required: true
        type: string
      - description: Password
        in: formData
        name: password
        required: true
        type: string
      - description: Password confirmation
        in: formData
        name: password2
        required: true
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: Rendered register form with errors
          schema:
            type: string
        "302":
          description: Redirect to login page
          schema:
            type: string
      summary: Register user
      tags:
      - Auth
  /api/search:
    get:
      description: Search stored pages and cached external results.
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      - description: Language code (default en)
        in: query
        name: language
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Search results
          schema:
            $ref: '#/definitions/handlers.APISearchResponse'
      summary: Search content
      tags:
      - Search
  /api/weather:
    get:
      description: Returns the current Copenhagen forecast used by the /weather page.
      produces:
      - application/json
      responses:
        "200":
          description: Forecast retrieved
          schema:
            $ref: '#/definitions/handlers.WeatherAPIResponse'
        "503":
          description: Weather service unavailable
          schema:
            $ref: '#/definitions/handlers.APIErrorResponse'
      summary: Get weather forecast
      tags:
      - API
  /healthz:
    get:
      description: Returns ok when the service is running.
      produces:
      - text/plain
      responses:
        "200":
          description: ok
          schema:
            type: string
        "500":
          description: internal error
          schema:
            type: string
      summary: Liveness probe
      tags:
      - Health
  /readyz:
    get:
      description: Checks database connectivity.
      produces:
      - text/plain
      responses:
        "200":
          description: ready
          schema:
            type: string
        "503":
          description: database not ready
          schema:
            type: string
      summary: Readiness probe
      tags:
      - Health
swagger: "2.0"